requirements.txt
git+https://github.com/openai/CLIP.git
git+https://github.com/facebookresearch/detectron2.git
einops==0.6.0
numpy==1.21.6
open_clip_torch==2.7.0
opencv_python==4.6.0.66
Pillow==9.3.0
pycocotools==2.0.2
scikit-image==0.19.3
scipy==1.7.3
setuptools==65.5.0
submitit==1.4.5
timm==0.5.4
torch==1.13.0
torchvision==0.14.0
tqdm==4.64.1
transformers==4.24.0
lvis==0.5.3
